refactor: refactor swap rate display into a separate component
